Package setting's view cells with Settings
